Add play to objkt view
