+++ components/views/styles.module.css
@@ -111,6 +111,10 @@
     font-size: 14px;
 }
 
+.playPause {
+    margin-right: 10px;
+}
+
 @media screen and (min-width: 600px) {
     .walletInput {
         max-width: 320px;
